pyproject.toml
[tool.poetry]
name = "django-calendar-app"
version = "0.1.0"
description = ""
authors = ["Tomas <tom.hanzely@gmail.com>"]
readme = "README.md"
packages = [{include = "django_calendar_app"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.7"
django-rest-framework = "^0.1.0"
django-timezone-field = "^5.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
